+++ builder/dockerfile/shell/wordsTest
